central/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-core==2.23.4
python-multipart==0.0.9


opencv-python==4.10.0.84
onnxruntime==1.23.2
numpy==1.26.4

pillow==11.0.0


requests==2.32.3

passlib[bcrypt]

python-jose

jinja2 

weasyprint